Moved waitUntilFinished into the shared URL service test utils
ref https://linear.app/ghost/issue/HKG-1919

The URL service parity suite carried its own copy of this helper and
the sitemap parity suite arriving behind this change needs the same
one. A second local copy is exactly the drift the shared utils file
exists to prevent, so the helper moves there first, with the existing
suite continuing to exercise it.
